requirements.txt
requests
dnspython
pytest
responses